requirements.txt
fastapi
uvicorn
langchain_openai
langchain_core
langchain
python-dotenv
pydantic
langchain-qdrant
qdrant-client
langchain-community
psycopg[binary]
psycopg-pool
langgraph
langgraph-checkpoint-postgres
langchain-nvidia-ai-endpoints